update location of v8 include path
in 4.5.x, v8 starts expecting the include paths to contain the root of
the v8 source directory, and everywhere it had previously referenced
`v8.h`, it now references `include/v8.h`. insted of 'v8-profiler.h', it
refererces 'include/v8-profiler.h', etc...

This adds a test to make sure it can find 'include/v8.h' in the include
path as part of the installation verification.