update to 8.4.371.19
This requires changing our checkout strategy to check out the branch-heads branch (as documented at https://v8.dev/docs/version-numbers) instead of using the 8.4.371 branch, which never gets patches (it's always locked to 8.4.371.0).

We could also specify the v8 patch version manually, and i'm open to doing that, just wanted to do this to get it working.